Created a new component (CustomRegisterForm), optmized validations and reverted RegisterForm
